# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: y-agent Web API
  version: 0.1.0
  description: |
    RESTful API for y-agent — exposes the same business logic as CLI/TUI.
    All endpoints return JSON. Errors follow RFC 7807 Problem Details.
  contact:
    name: y-agent team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/v1/diagnostics/traces:
    get:
      operationId: listTraces
      tags: [Diagnostics]
      summary: List recent traces
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
          description: Filter by session ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of traces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TraceInfo"
  /api/v1/diagnostics/traces/{trace_id}:
    get:
      operationId: getTrace
      tags: [Diagnostics]
      summary: Get a trace with observations
      parameters:
        - name: trace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Trace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceDetail"
        "404":
          $ref: "#/components/responses/NotFound"

# ═══════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code (e.g. "not_found", "bad_request")
        message:
          type: string
          description: Human-readable error message
...
    TraceInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        started_at:
          type: string
          format: date-time
        total_cost_usd:
          type: number
        total_input_tokens:
          type: integer
        total_output_tokens:
          type: integer
        total_duration_ms:
          type: integer
          nullable: true
    TraceDetail:
      type: object
      properties:
        trace:
          $ref: "#/components/schemas/TraceInfo"
        observations:
          type: array
          items:
            type: object